URGENT FIX: Buffer overflow in PATHOPS_ERROR_INFO
Fixed critical heap buffer overflow in Win32IDE_Sidebar_PathOps.cpp:

CHANGE:
- Increased PATHOPS_ERROR_INFO.message buffer from 512 to 8192 wchar_t
- Updated FormatMessageW call from 512 to 8192 buffer size

ROOT CAUSE:
When PathOps_GetLastErrorInfo() was called during IDE startup, the fixed 512-character
buffer in the heap-allocated PATHOPS_ERROR_INFO struct wasn't large enough to hold
system error messages from FormatMessageW(). This caused the CRT to detect
'application wrote to memory after end of head buffer' error.

TESTED:
- Build succeeds without errors
- IDE launches and runs without buffer overflow crash
- Process runs stably at 84MB memory

This resolves the immediate heap corruption issue (-0xC0000374).

+++ src/core/crash_containment.cpp
@@ -147,7 +147,7 @@ static bool writeMiniDump(const char* path, EXCEPTION_POINTERS* ep, DWORD dumpTy
     mei.ClientPointers = FALSE;
 
     BOOL ok = pWriteDump(GetCurrentProcess(), GetCurrentProcessId(),
-                         hFile, dumpType, &mei, nullptr, nullptr);
+                         hFile, (DWORD)dumpType, &mei, nullptr, nullptr);
 
     CloseHandle(hFile);
     FreeLibrary(hDbgHelp);
